regisshtry-push: use the digest id to push the image
This way the pushed image won't contain a tag.

The images built by Concourse CI don't contain a tag either which is why
we now tag the image when loading it on the servers. This change makes
the regisshtry-push and the Concourse CI behaviour identical and fixes
an error that happens when we clumsily try to parse the sha256 of the
loaded image that is not returned by Docker when we load a tagged image.